ACPI: scan: Turn off unused power resources during initialization
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1982319
Upstream: Linux v5.13
Tested by: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=38078369
            x86_64 (debug) software raid fails tests, but it does on
            current main

commit 4b9ee772eaa82188b0eb8e05bdd1707c2a992004
Author: "Rafael J. Wysocki" <rafael.j.wysocki@intel.com>
Date: Thu, 18 Mar 2021 19:25:12 +0100

It is reported that on certain platforms there are power resources
that are not associated with any devices physically present in the
platform.  Those power resources are expected to be turned off by
the OS in accordance with the ACPI specification (section 7.3 of
ACPI 6.4) which currently is not done by Linux and that may lead
to obscure issues.

For instance, leaving those power resources in the "on" state may
prevent the platform from reaching the lowest power state in
suspend-to-idle which leads to excessive power draw.

For this reason, turn all of the unused ACPI power resources off
at the end of the initial namespace scan for devices in analogy with
resume from suspend-to-RAM.

Link: https://uefi.org/specs/ACPI/6.4/07_Power_and_Performance_Mgmt/device-power-management-objects.html
Reported-by: David Box <david.e.box@linux.intel.com>
Tested-by: Wendy Wang <wendy.wang@intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit 4b9ee772eaa82188b0eb8e05bdd1707c2a992004)

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>