igb: handle vlan types with checker enabled
Author: Jesse Brandeburg <jesse.brandeburg@intel.com>

The sparse build (C=2) finds some issues with how the driver
dealt with the (very difficult) hardware that in some generations
uses little-endian, and in others uses big endian, for the VLAN
field. The code as written picks __le16 as a type and for some
hardware revisions we override it to __be16 as done in this
patch. This impacted the VF driver as well so fix it there too.

Also change the vlan_tci assignment to override the sparse
warning without changing functionality.

Signed-off-by: Jesse Brandeburg <jesse.brandeburg@intel.com>
Tested-by: Dave Switzer <david.switzer@intel.com>
Signed-off-by: Tony Nguyen <anthony.l.nguyen@intel.com>
(cherry picked from commit c7cbfb028b95360403d579c47aaaeef1ff140964)
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1970662
Signed-off-by: Corinna Vinschen <vinschen@redhat.com>