perf/arm-ccn: Use irq_set_affinity()
Upstream Status: Linus v5.14
Testing Status: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=40498166
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1997732

commit 84fca8ba620581067c16f2b578f277b1c72fb74b
Author: Thomas Gleixner <tglx@linutronix.de>
Date: Tue, 18 May 2021 11:17:27 +0200

The driver uses irq_set_affinity_hint() to set the affinity for the PMU
interrupts, which relies on the undocumented side effect that this function
actually sets the affinity under the hood.

Setting an hint is clearly not a guarantee and for these PMU interrupts an
affinity hint, which is supposed to guide userspace for setting affinity,
is beyond pointless, because the affinity of these interrupts cannot be
modified from user space.

Aside of that the error checks are bogus because the only error which is
returned from irq_set_affinity_hint() is when there is no irq descriptor
for the interrupt number, but not when the affinity set fails. That's on
purpose because the hint can point to an offline CPU.

Replace the mindless abuse with irq_set_affinity().

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Cc: Will Deacon <will@kernel.org>
Cc: Mark Rutland <mark.rutland@arm.com>
Cc: linux-arm-kernel@lists.infradead.org
Acked-by: Mark Rutland <mark.rutland@arm.com>
Link: https://lore.kernel.org/r/20210518093118.128250213@linutronix.de
Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit 84fca8ba620581067c16f2b578f277b1c72fb74b)

Signed-off-by: Mark Langsdorf <mlangsdo@redhat.com>