cifs: refactor create_sd_buf() and and avoid corrupting the buffer
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1812436
commit ea64370bcae126a88cd26a16f1abcc23ab2b9a55

When mounting with "idsfromsid" mount option, Azure
corrupted the owner SIDs due to excessive padding
caused by placing the owner fields at the end of the
security descriptor on create.  Placing owners at the
front of the security descriptor (rather than the end)
is also safer, as the number of ACEs (that follow it)
are variable.

Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>
    Suggested-by: Rohith Surabattula <rohiths@microsoft.com>
    CC: Stable <stable@vger.kernel.org> # v5.8
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit ea64370bcae126a88cd26a16f1abcc23ab2b9a55)
    Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>