vp_vdpa: allow set vq state to initial state after reset
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1975173

We used to fail the set_vq_state() since it was not supported yet by
the virtio spec. But if the bus tries to set the state which is equal
to the device initial state after reset, we can let it go.

This is a must for virtio_vdpa() to set vq state during probe which is
required for some vDPA parents.

Signed-off-by: Jason Wang <jasowang@redhat.com>
Link: https://lore.kernel.org/r/20210602021536.39525-4-jasowang@redhat.com
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Reviewed-by: Eli Cohen <elic@nvidia.com>
(cherry picked from commit 1225c216d9542ed4883027d3af50035c35a7a03c)
Signed-off-by: Cindy Lu <lulu@redhat.com>