perf pmu: Save pmu name
Bugzilla: https://bugzilla.redhat.com/1838635

upstream
========
commit 32705de7d45d0ed989517a63454c2b3e5e5ea267
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Tue Apr 27 15:01:17 2021 +0800

description
===========
On hybrid platform, one event is available on one pmu
(such as, available on cpu_core or on cpu_atom).

This patch saves the pmu name to the pmu field of struct perf_pmu_alias.
Then next we can know the pmu which the event can be enabled on.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>