perf/x86/intel/uncore: Rename uncore_notifier to uncore_pci_sub_notifier
Bugzilla: https://bugzilla.redhat.com/1837330

upstream
========
commit 6477dc3934775f82a571fac469fd8c348e611095
Author: Kan Liang <kan.liang@linux.intel.com>
Date: Wed Mar 17 10:59:35 2021 -0700

description
===========
Perf will use a similar method to the PCI sub driver to register
the PMUs for the PCI type of uncore blocks. The method requires a BUS
notifier to support hotplug. The current BUS notifier cannot be reused,
because it searches a const id_table for the corresponding registered
PMU. The PCI type of uncore blocks in the discovery tables doesn't
provide an id_table.

Factor out uncore_bus_notify() and add the pointer of an id_table as a
parameter. The uncore_bus_notify() will be reused in the following
patch.

The current BUS notifier is only used by the PCI sub driver. Its name is
too generic. Rename it to uncore_pci_sub_notifier, which is specific for
the PCI sub driver.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>