bnxt_en: Send HWRM_FUNC_RESET fw command unconditionally.
In the AER or firmware reset flow, if we are in fatal error state or
if pci_channel_offline() is true, we don't send any commands to the
firmware because the commands will likely not reach the firmware and
most commands don't matter much because the firmware is likely to be
reset imminently.

However, the HWRM_FUNC_RESET command is different and we should always
attempt to send it.  In the AER flow for example, the .slot_reset()
call will trigger this fw command and we need to try to send it to
effect the proper reset.

Fixes: b340dc680ed4 ("bnxt_en: Avoid sending firmware messages when AER error is detected.")
Reviewed-by: Edwin Peer <edwin.peer@broadcom.com>
Signed-off-by: Vasundhara Volam <vasundhara-v.volam@broadcom.com>
Signed-off-by: Michael Chan <michael.chan@broadcom.com>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 825741b071722f1c8ad692cead562c4b5f5eaa93)
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1921449
Build Info: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=38028096
Tested: compile tested
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>