cifs: fix fallocate when trying to allocate a hole.
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1866684
Tested: Manual testing and upstream buildbot

commit 488968a8945c119859d91bb6a8dc13bf50002f15

Remove the conditional checking for out_data_len and skipping the fallocate
if it is 0. This is wrong will actually change any legitimate the fallocate
where the entire region is unallocated into a no-op.

Additionally, before allocating the range, if FALLOC_FL_KEEP_SIZE is set then
we need to clamp the length of the fallocate region as to not extend the size of the file.

Fixes: 966a3cb7c7db ("cifs: improve fallocate emulation")
    Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>
    Signed-off-by: Steve French <stfrench@microsoft.com>
    (cherry picked from commit 488968a8945c119859d91bb6a8dc13bf50002f15)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>