bpf: bpftool: Add -fno-asynchronous-unwind-tables to BPF Clang invocation
Upstream Status: RHEL-only
Bugzilla: https://bugzilla.redhat.com/1997124

llvm12 still produces .eh_frame section for bpftool internal progs
which causes warnings:

  libbpf: elf: skipping unrecognized data section(32) .eh_frame
  libbpf: elf: skipping relo section(33) .rel.eh_frame for section(32) .eh_frame

The same was fixed upstream for samples/bpf with

5a0ae9872d5c ("bpf, samples: Add -fno-asynchronous-unwind-tables to BPF Clang invocation")

Upstream updated llvm that it does not produce the sections now, so
the change most likely will not be welcomed there, and RHEL deadline
is too close.

The same applied for bpf tests, but it's not important since it's
not seen to customers.

Signed-off-by: Yauheni Kaliuta <ykaliuta@redhat.com>